added test cases, solved lint errors

+++ test/test.js
@@ -2,7 +2,8 @@ const chai = require("chai");
 const _helpers = require("../lib/helpers");
 const chaiHttp = require("chai-http");
 const app = require("../index");
-
+const fs = require("fs");
+const path = require("path");
 chai.should();
 // Configure chai
 
@@ -22,15 +23,20 @@ describe("/GET owners", () => {
   });
 });
 
+/* negative test case - trying to get all pets when the file is not present*/
 describe("Pets", () => {
   beforeEach(done => {
     //Before each test remove the pets.json file
-    _helpers
-      .removeFile("pets.json")
-      .then(resp => {
-        done();
-      })
-      .catch(done);
+    if (fs.existsSync(path.join(__dirname + "/../data/pets.json"))) {
+      _helpers
+        .removeFile("pets.json")
+        .then(resp => {
+          done();
+        })
+        .catch(done);
+    } else {
+      done();
+    }
   });
 
   // test the GET route of pets when the file is not present
@@ -47,11 +53,13 @@ describe("Pets", () => {
   });
 });
 
+/* POST the data to test positive test case of get /pets */
 //POST route of pets
+
 describe("/POST pets", () => {
   it("it should POST a pet", done => {
     const pet = {
-      name: "testPet",
+      name: "updatingPet",
       color: "red",
       breed: "lab",
       age: "12",
@@ -69,7 +77,29 @@ describe("/POST pets", () => {
       });
   });
 });
+// adding two sets of pets. This is to ensure PUT is working fine.
+describe("/POST pets", () => {
+  it("it should POST a pet", done => {
+    const pet = {
+      name: "testPet",
+      color: "red",
+      breed: "lab",
+      age: "12",
+      type: "dog",
+      ownedBy: "1"
+    };
 
+    chai
+      .request(app)
+      .post("/pets")
+      .send(pet)
+      .end((err, res) => {
+        res.should.have.status(200);
+        res.body.should.be.a("object");
+        done();
+      });
+  });
+});
 //get all pets
 describe("/GET pets", () => {
   it("it should GET all the pets", done => {
@@ -84,6 +114,9 @@ describe("/GET pets", () => {
   });
 });
 
+/* negative test cases of Post route
+ testing y removing each required field
+*/
 describe("/POST pets", () => {
   it("it should not POST a pet without name field", done => {
     const pet = {
@@ -146,6 +179,50 @@ describe("/POST pets", () => {
       });
   });
 });
+/* Restricted age of pets between 1 and 50 */
+describe("/POST pets", () => {
+  it("it should not POST a pet with age field value less than 0", done => {
+    const pet = {
+      name: "test1",
+      color: "red",
+      breed: "lab",
+      type: "dog",
+      pet: "0",
+      ownedBy: "1"
+    };
+    chai
+      .request(app)
+      .post("/pets")
+      .send(pet)
+      .end((err, res) => {
+        res.should.have.status(422);
+        res.body.should.be.a("object");
+        done();
+      });
+  });
+});
+
+describe("/POST pets", () => {
+  it("it should not POST a pet with age field value greater than 50 ", done => {
+    const pet = {
+      name: "test1",
+      color: "red",
+      breed: "lab",
+      type: "dog",
+      pet: "51",
+      ownedBy: "1"
+    };
+    chai
+      .request(app)
+      .post("/pets")
+      .send(pet)
+      .end((err, res) => {
+        res.should.have.status(422);
+        res.body.should.be.a("object");
+        done();
+      });
+  });
+});
 describe("/POST pets", () => {
   it("it should not POST a pet without breed field", done => {
     const pet = {
@@ -187,6 +264,7 @@ describe("/POST pets", () => {
   });
 });
 
+/* type of pet must be cat or dog */
 describe("/POST pets", () => {
   it("it should not POST a pet if type is neither cat nor dog field", done => {
     const pet = {
@@ -229,6 +307,7 @@ describe("/POST pets", () => {
   });
 });
 
+/* Get pets belonging to particular owner */
 describe("/GET/:id owner", () => {
   it("it should GET a pet by the given owner id", done => {
     const pet = {
@@ -242,7 +321,7 @@ describe("/GET/:id owner", () => {
 
     chai
       .request(app)
-      .get("/owner/" + pet.ownedBy)
+      .get(`/owner/${pet.ownedBy}`)
       .send(pet)
       .end((err, res) => {
         res.should.have.status(200);
@@ -252,6 +331,7 @@ describe("/GET/:id owner", () => {
   });
 });
 
+/* Negative test case where owner is not present or owner does not own any pet  */
 describe("/GET/:id owner", () => {
   it("it should not GET a pet by the given owner id", done => {
     const pet = {
@@ -265,7 +345,7 @@ describe("/GET/:id owner", () => {
 
     chai
       .request(app)
-      .get("/owner/" + pet.ownedBy)
+      .get(`/owner/${pet.ownedBy}`)
       .send(pet)
       .end((err, res) => {
         res.should.have.status(200);
@@ -291,7 +371,7 @@ describe("/PUT/:id pet", () => {
 
     chai
       .request(app)
-      .put("/pets/" + pet.id)
+      .put(`/pets/${pet.id}`)
       .send({
         pet
       })
@@ -314,7 +394,7 @@ describe("/PUT/:id pets", () => {
     };
     chai
       .request(app)
-      .put("/pets/" + pet.id)
+      .put(`/pets/${pet.id}`)
       .send(pet)
       .end((err, res) => {
         res.should.have.status(422);
@@ -336,7 +416,7 @@ describe("/PUT/:id pets", () => {
     };
     chai
       .request(app)
-      .put("/pets/" + pet.id)
+      .put(`/pets/${pet.id}`)
       .send(pet)
       .end((err, res) => {
         res.should.have.status(422);
@@ -357,7 +437,7 @@ describe("/PUT/:id pets", () => {
     };
     chai
       .request(app)
-      .put("/pets/" + pet.id)
+      .put(`/pets/${pet.id}`)
       .send(pet)
       .end((err, res) => {
         res.should.have.status(422);
@@ -378,7 +458,7 @@ describe("/PUT/:id pets", () => {
     };
     chai
       .request(app)
-      .put("/pets/" + pet.id)
+      .put(`/pets/${pet.id}`)
       .send(pet)
       .end((err, res) => {
         res.should.have.status(422);
@@ -399,7 +479,7 @@ describe("/PUT/:id pets", () => {
     };
     chai
       .request(app)
-      .put("/pets/" + pet.id)
+      .put(`/pets/${pet.id}`)
       .send(pet)
       .end((err, res) => {
         res.should.have.status(422);
@@ -422,7 +502,7 @@ describe("/PUT/:id pets", () => {
     };
     chai
       .request(app)
-      .put("/pets/" + pet.id)
+      .put(`/pets/${pet.id}`)
       .send(pet)
       .end((err, res) => {
         res.should.have.status(422);
@@ -434,6 +514,7 @@ describe("/PUT/:id pets", () => {
 describe("/PUT/:id pets", () => {
   it("it should not update a pet without ownedBy field", done => {
     const pet = {
+      id: "1",
       name: "test1",
       breed: "lab",
       age: "12",
@@ -442,12 +523,88 @@ describe("/PUT/:id pets", () => {
     };
     chai
       .request(app)
-      .put("/pets/" + pet.id)
+      .put(`/pets/${pet.id}`)
       .send(pet)
       .end((err, res) => {
         res.should.have.status(422);
         res.body.should.be.a("object");
         done();
       });
+  });
+});
+describe("/PUT/:id pets", () => {
+  it("it should not update a pet with age field value less than or equal to 0", done => {
+    const pet = {
+      id: "1",
+      name: "test1",
+      breed: "lab",
+      age: "0",
+      type: "dog",
+      color: "red"
+    };
+    chai
+      .request(app)
+      .put(`/pets/${pet.id}`)
+      .send(pet)
+      .end((err, res) => {
+        res.should.have.status(422);
+        res.body.should.be.a("object");
+        done();
+      });
+  });
+});
+
+describe("/PUT/:id pets", () => {
+  it("it should not update a pet with age field value greater than 50 ", done => {
+    const pet = {
+      id: "1",
+      name: "test1",
+      color: "red",
+      breed: "lab",
+      type: "dog",
+      age: "51",
+      ownedBy: "1"
+    };
+    chai
+      .request(app)
+      .put(`/pets/${pet.id}`)
+      .send(pet)
+      .end((err, res) => {
+        res.should.have.status(422);
+        res.body.should.be.a("object");
+        done();
+      });
+  });
+});
+
+/* success case of updating */
+describe("/PUT/:id pets", () => {
+  it("it should  update a pet ", done => {
+    let id = "";
+    _helpers
+      .read("pets.json")
+      .then(response => {
+        const pets = _helpers.stringToJson(response);
+        id = pets.pets[0].id;
+        const pet = {
+          id: id,
+          name: "updatedName",
+          color: "updatedColor",
+          breed: "updatedBreed",
+          type: "dog",
+          age: "10",
+          ownedBy: "1"
+        };
+        chai
+          .request(app)
+          .put(`/pets/${pet.id}`)
+          .send(pet)
+          .end((err, res) => {
+            res.should.have.status(200);
+            res.body.should.be.a("object");
+            done();
+          });
+      })
+      .catch(err => {});
   });
 });
